Initialize this repository
This commit is used to initialize this repository.

I didn't set up this Git repository when I started this project because
I forgot about it (probably because of my busy schoolwork). But now I
think it's time to set it up.

(By the way, it's the first time for me to write a pretty Git log. I
think I think I don't understand Git very well before.)

Signed-off-by: xianfengting <xianfengting221@163.com>